moves ai cli code to dlthub
