+++ dlt/_workspace/mcp/tools/toolkit_tools.py
@@ -4,7 +4,7 @@ from pydantic import Field
 from fastmcp.exceptions import ToolError
 
 from dlt.common.typing import Annotated
-from dlt._workspace.cli.ai.utils import (
+from dlt._workspace.cli.dlthub.ai.utils import (
     DEFAULT_AI_WORKBENCH_BRANCH,
     DEFAULT_AI_WORKBENCH_REPO,
     fetch_workbench_base,
