+++ tests/workspace/cli/ai/utils.py
@@ -10,8 +10,8 @@ import pytest
 import yaml
 
 from dlt.common.libs import git
-from dlt._workspace.cli.ai.agents import AI_AGENTS
-from dlt._workspace.cli.ai.utils import (
+from dlt._workspace.cli.dlthub.ai.agents import AI_AGENTS
+from dlt._workspace.cli.dlthub.ai.utils import (
     DEFAULT_AI_WORKBENCH_BRANCH,
     DEFAULT_AI_WORKBENCH_REPO,
     compute_file_hash,
